add user schema

+++ routes/heroes.js
@@ -1,8 +1,10 @@
 const express = require('express');
+const jwt = require('jsonwebtoken');
 const Hero = require('../models/hero');
 const router = express.Router();
 
 let heroArray = [{id: 1, name: 'Captain America'}, {id: 2, name: 'Iron Man'}, {id: 3, name: 'Black Widow'}]
+let SECRET_KEY = "1234567"
 
 router.get('/', async (req, res) => {
     // let heroes = ['Captain America', 'Iron Man', 'Black Widow'];
@@ -10,7 +12,7 @@ router.get('/', async (req, res) => {
 
     // let heroes = await Hero.find();
     // let heroes = await Hero.find({deceased:true});
-    let heroes = await Hero.find({ likeCount:{$gt:0} });
+    let heroes = await Hero.find({likeCount: {$gt: 0}});
     // let heroes = await Hero.find({ likeCount:{$eq:null} });
     // let heroes = await Hero.find({deceased: true}).select({name:1}); //select only name
     // let heroes = await Hero.find().sort({name:'asc'});
@@ -45,8 +47,19 @@ router.get('/:heroId', async (req, res) => {
 
 router.post('/', async (req, res) => {
 
-    if (!req.body.heroName)
+    const token = req.header("x-jwt-token");
+
+    if (!token) return res.status(401).send("Access denied. No token.");
+
+    try {
+        jwt.verify(token, SECRET_KEY);
+    } catch (e) {
+        res.status(400).send("Invalid token")
+    }
+
+    if (!req.body.heroName) {
         return res.status(400).send("Please add mandotory data");
+    }
 
     try {
         let heroToBeAddedToDb = new Hero({
@@ -97,9 +110,9 @@ router.post('/', async (req, res) => {
 router.put('/:heroId', async (req, res) => {
 
     let hero = await Hero.findByIdAndUpdate(
-        {_id:req.params.heroId},
+        {_id: req.params.heroId},
         {$set: {name: req.body.heroName}},
-        {new:true, useFindAndModify:false}
+        {new: true, useFindAndModify: false}
     );
     res.send(hero);
 
@@ -119,20 +132,36 @@ router.put('/:heroId', async (req, res) => {
 //     res.send(heroArray);
 // });
 
-router.delete('/:heroId', async (req,res) =>{
+router.delete('/:heroId', async (req, res) => {
     // let heroId = parseInt(req.params.heroId);
     // let hero = heroesArray.find(h=> h.id === heroId);
 
-    let hero = await Hero.findById(req.params.heroId);
 
-    if(!hero){
+    const token = req.header("x-jwt-token");
+
+    if (!token) return res.status(401).send("Access denied. No token.");
+
+    try {
+        jwt.verify(token, SECRET_KEY);
+    } catch (e) {
+        res.status(400).send("Invalid token")
+    }
+
+    let decoded = jwt.decode(token, SECRET_KEY);
+    if (!decoded.isAdmin) {
+        return res.status(403).send("No Permission to Delete");
+    }
+
+    let hero = await Hero.findById(req.params.heroId);
+    if (!hero) {
         return res.status(404).send("The given Id does not our server ");
     }
+
 
     // let deleteHero = heroesArray.indexOf(hero);
     // heroesArray.splice(deleteHero,1);
 
-    hero.deleteOne({_id :req.params.heroId });
+    hero.deleteOne({_id: req.params.heroId});
 
     res.send(hero);
 })
